% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{addBarSegs}
\alias{addBarSegs}
\title{Title}
\usage{
addBarSegs(gr.chr, gr, col.ids = c(1), adj.x = 0,
  cn.colors = c("black"), cn.range = NULL, chr.lines = FALSE,
  yrange = c(0, 1), y.spacer = 0.1)
}
\arguments{
\item{gr.chr}{}

\item{gr}{}

\item{col.ids}{}

\item{adj.x}{}

\item{cn.colors}{}

\item{cn.range}{}

\item{chr.lines}{}

\item{yrange}{}

\item{y.spacer}{}
}
\description{
Title
}
