% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{addCnSegs}
\alias{addCnSegs}
\title{GR Viz: Plotting function to add copy-number segments}
\usage{
addCnSegs(gr.chr, col.ids = c(1), adj.x = 0, cn.colors = c("blue",
  "white", "red"), cn.range = NULL, chr.lines = FALSE, yrange = c(0, 1),
  y.spacer = 0.1)
}
\arguments{
\item{gr.chr}{[GRanges]: Copy-number GRanges object with CN data in elementMetadata}

\item{col.ids}{[Vector]: Which columns/samples to plot and in what order}

\item{adj.x}{[Integer]: Pulled from plot.settings to adjust the x-coordinates for the right chromosome}

\item{cn.colors}{[Char. Vector]: Range of colours to feed into ColorRampPalette [Default: c("blue", "white", "red")]}

\item{cn.range}{[Num. Vector]: Numeric vector of colour ranges [Default: seq(-3, 3, by=0.5)]}

\item{chr.lines}{[Boolean]: Whether to add dashed grey lines indicating edges of chromosomes}

\item{yrange}{[Int. Vector]: The ylim plotting space}

\item{y.spacer}{[Numeric]: How much space to add between tracks}
}
\description{
Works in tandem with plotGrMetadata() to add coloured CN rectangles to the existing plot
}
