% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{addExprScores}
\alias{addExprScores}
\title{GR Viz: Plotting function to add Expression ZScores}
\usage{
addExprScores(gr.chr, yrange, adj.x, anno.track = 0.25, add.y.axis = FALSE,
  add.annotations = TRUE, min.anno.gap = 5e+07, add.anno.lines = TRUE,
  ...)
}
\arguments{
\item{gr.chr}{[GRanges]: Input GRanges.  Contains elementMEtadata() columns of "zscore" and "symbol"}

\item{yrange}{[Int. Vector]: The y-axis range to plot on; i.e. c(0, 5)}

\item{adj.x}{[Integer]: Pulled from plot.settings to adjust the x-coordinates for the right chromosome}

\item{anno.track}{[Numeric]: The fraction of the plot to allocate for an extra annotation track}

\item{add.y.axis}{[Boolean]: Whether to add an y-axis with z-score values}

\item{add.annotations}{[Boolean]: Whether to add annotations for each zscore range}
}
\description{
Still in the developmental build: Adds zscores of matrix for a given single sample where elementMetadata contains both symbol for Gene HUGO symbol and zscore columns
}
