% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{aggregateGr}
\alias{aggregateGr}
\title{cnTools: Aggregate Genomic Ranges}
\usage{
aggregateGr(list.gr)
}
\arguments{
\item{list.gr}{[GRangesList]: a list of granges objects or GRangesList}
}
\value{
A single GRanges object with each row of the elementMetadata() containing all the samples copy-number values for that segment
}
\description{
Because different samples have different segmentations, this function attempts to create a unified matrix with the fewest number of segments needed to represent all copy-number segments from all samples.  This allows for the elementMetadata() to store the associate copy-number values for those samples all in one matrix.
}
\examples{
aggregateGr(list(gr1, gr2))
}
