% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{assignAmpDel}
\alias{assignAmpDel}
\title{cnTools: Convert Log2Ratio to Amp/Del}
\usage{
assignAmpDel(seg.gr, cn.thresh = 0.5)
}
\arguments{
\item{[GRanges]:}{seg.gr A GRanges object with copy-number log2ratios in elementMetadata() columns}

\item{[Integer]:}{cn.thresh A log2ratio cutoff to indicate gain or loss}
}
\value{

}
\description{
A temporary helper function to truncate log2ratios at 0.5 and turn them into Amp (1) or Del (-1)
}
