% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonetrack.R
\name{cdr3_dataframe.fx}
\alias{cdr3_dataframe.fx}
\title{Dataframe for clonetracking}
\usage{
cdr3_dataframe.fx(datapath, chain, filelist, totalinframe)
}
\arguments{
\item{datapath}{path to mixcr files}

\item{chain}{any of: TRA, TRB, TRD, TRG, IGH, IGL, IGK}

\item{filelist}{a list of files to track clonotypes. Output from list.files()}

\item{totalinframe}{"total" outputs all CDR3s, "inframe" removes non-productive CDR3s}
}
\description{
This function takes a list of files and compiles them in one dataframe
for use in plot_clonetracks.fx. Also you can choose to exclude non-productive clonotypes
}
