% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{chooseTranscript}
\alias{chooseTranscript}
\title{chooseTranscript}
\usage{
chooseTranscript(edb, gene, Chromosome, Start_Position, End_Position,
  default.longest = TRUE)
}
\arguments{
\item{edb}{Ensembl DB R package, v86=GRCh38, v75=GRCh37 (Default: EnsDb.Hsapiens.v86)}

\item{gene}{Gene of interest in HUGO format (e.g., MTOR). If present, Locus infomation is not required.}

\item{Chromosome}{Chromosome}

\item{Start_Position}{Mutation start position}

\item{End_Position}{Mutation end position}

\item{default.longest}{Selects the longest ENST txid as the canonical txid (Default: TRUE)}
}
\description{
Select a single transcript using a gene name or mutation locus selection method
}
\examples{
chooseTranscript(EnsDb.Hsapiens.v86,"MTOR")
}
