% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{cleanSeg}
\alias{cleanSeg}
\title{Cleans the segfile based on the type of input. "Cleaning" will depend on data.type}
\usage{
cleanSeg(seg, data.type = "ichor")
}
\arguments{
\item{seg}{A segfile read in as a dataframe}

\item{data.type}{[default='ichor']}
}
\description{
Cleans the segfile based on the type of input. "Cleaning" will depend on data.type
}
