% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCR_tools.R
\name{clontrack.fx}
\alias{clontrack.fx}
\title{Hello people!}
\usage{
clontrack.fx(datapath, plotpath, patient_id, cycleorder, countfrac)
}
\arguments{
\item{datapath}{path to mixcr TRB clones}

\item{plotpath}{path to plot directory}

\item{patient_id}{in this format: INS-X-XXX}

\item{cycleorder}{order the plot. In this format. c("SB", "C3B", "C6B", etc)}

\item{countfrac}{either cloneCount or cloneFraction}
}
\description{
This clonetrack function is for INSPIRE project. Enjoy!
}
\examples{


}
