% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{cnDist}
\alias{cnDist}
\title{cnTools: Calculate distances between CN-profiles}
\usage{
cnDist(gr, method = "euclidean")
}
\arguments{
\item{gr}{GRanges object}

\item{method}{'euclidean', 'mhamming'=modified hamming, 'pearson'}
}
\description{
cnTools: Calculate distances between CN-profiles
}
