% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{collapseSample}
\alias{collapseSample}
\title{cnTools: Wrapper for copy-number metrics}
\usage{
collapseSample(gr, sample.idx, na.rm = TRUE, continuous.intervals = TRUE)
}
\arguments{
\item{gr}{[GRanges]: GRanges object with copy-number in elementMetadata()}

\item{sample.idx}{[Integer]: the sample index in the metadata}

\item{na.rm}{[Boolean]: Remove NA or keep NAs in the reduction}
}
\description{
Takes an individual sample and reduces GRanges ranges where copy-number does not change between subsequent ranges
}
