% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataframeToGranges}
\alias{dataframeToGranges}
\title{utils: dataframeToGranges}
\usage{
dataframeToGranges(intdf, chr.regex = "^chr(om|omosome)?$",
  start.regex = "^(loc|seg)?(.|_)?start(s|_idx)?$",
  end.regex = "^(loc|seg)?(.|_)?end(s|_idx)?$", ...)
}
\arguments{
\item{intdf}{An dataframe containing segment intervals}

\item{chr.regex}{Regex for the chromosome column}

\item{start.regex}{Regex for the start index column}

\item{end.regex}{Regex for the end index column}

\item{...}{}
}
\description{
OUTDATED: Uses a set of complex regex expression to automatically parse the chromosome, start_idx, and end_idx from a dataframe of segments to create a GRanges object. Also accepts standard GRanges arguments that you would normally pass in to define extra information.
}
