% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{demo.plotLikelihoodRatio}
\alias{demo.plotLikelihoodRatio}
\title{Vis.demo: demo.plotLikelihoodRatio}
\usage{
demo.plotLikelihoodRatio()
}
\description{
An example piece of code to run the plotLikelihoodRatio() function
}
\examples{
demo.plotLikelihoodRatio()
}
