% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{demo.plotScatterLine}
\alias{demo.plotScatterLine}
\title{Vis.demo: demo.plotScatterLine}
\usage{
demo.plotScatterLine()
}
\description{
An example piece of code to run the plotScatterLine() function
}
\examples{
demo.plotScatterLine()
}
