% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune_Diversity.R
\name{Divstats.fx}
\alias{Divstats.fx}
\title{Matrix for diversity measures and related descriptives}
\usage{
Divstats.fx(lst, chain, batchname, outpath)
}
\arguments{
\item{lst}{list output from immunelistfx}

\item{chain}{any of: TRA, TRB, TRD, TRG, IGH, IGL, IGK}

\item{batchname}{output csv filename will be in this form: "divstats_chain_batchname.csv"}

\item{outpath}{path to save output file}
}
\description{
This function takes the list generated from immunelistfx and generates a dataframe for descriptives and diversity estimators
}
\examples{
data(celllines_list)
Divstats.fx(celllines_list, "TRB", "thisismytest", "~/")

}
