% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{.getDomainBimap}
\alias{.getDomainBimap}
\title{.getDomainBimap}
\usage{
.getDomainBimap(domain.src)
}
\arguments{
\item{domain.src}{}
}
\value{

}
\description{
Maps protein feature IDs to protein feature domain names
 (i.e.  PF02260 -> FATC)
}
\examples{
.getDomainBimap('pfam')
}
