% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genDemoData}
\alias{genDemoData}
\title{utils: genDemoData}
\usage{
genDemoData(seed = 404)
}
\arguments{
\item{seed}{Set the random seed [default: 404]}
}
\description{
Generates a random granges object to be used for demo purposes when trying out the cnSignature functions
}
\examples{
demo <- genDemoData(404)
}
