% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportal_tools.R
\name{genMEMatrix}
\alias{genMEMatrix}
\title{Genreates Mutual exclusivitiy matrices}
\usage{
genMEMatrix(tbl, val = "Log2 Odds Ratio")
}
\arguments{
\item{tbl}{The table.tsv downloaded from cBioportal Mutual Exclusivity
tab and imported via read.table with headers}

\item{val}{Value that you want to encode in the matrix}
}
\value{
Returns a matrix of A by B for given Val
}
\description{
Takes the table.tsv downloadable file from cBioportal
and outputs a matrix of values for A and B columns
}
\examples{
or.mat <- genMEMatrix(tbl, val='Log2 Odds Ratio')
}
