% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{getCdsDat}
\alias{getCdsDat}
\title{getCdsDat}
\usage{
getCdsDat(txid, edb, gdb)
}
\arguments{
\item{txid}{Transcript ID in Ensembl format}

\item{edb}{Ensembl DB R package, v86=GRCh38, v75=GRCh37 (Default: EnsDb.Hsapiens.v86)}

\item{gdb}{genome sequence database (Default: BSgenome.Hsapiens.UCSC.hg38)}
}
\description{
For a given transcript, return the coding sequence
}
\examples{
getCdsDat("ENST00000471181", EnsDb.Hsapiens.v86, BSgenome.Hsapiens.UCSC.hg38)
}
