% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{getGieCol}
\alias{getGieCol}
\title{GR Viz: Generates the colours for cytobands}
\usage{
getGieCol(gie, gie.col = "black", cen.col = "red", plot.cband = TRUE,
  alpha.factor = 2)
}
\arguments{
\item{gie}{[Char. Vector]: A vector containing all the ordered UCSC cytoband Giemsa colour}

\item{gie.col}{[Character]: Colour for giemsa positive segments}

\item{cen.col}{[Character]: Colour for centromere segments}

\item{plot.cband}{[Boolean]: Toggle for plotting cytobands or not; will always plot centromeres}

\item{alpha.factor}{[Integer]: How many times to dilute the intensity of the colours, i.e. 2 will dilute the colours 2x [Default: 2]}
}
\description{
GR Viz: Generates the colours for cytobands
}
