% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMode}
\alias{getMode}
\title{utils: Calculates the mode}
\usage{
getMode(n, na.rm = FALSE)
}
\arguments{
\item{n}{[Vector]: A vector of numbers elements to calculate the mode}
}
\description{
utils: Calculates the mode
}
\examples{
getMode(c(rep(1,5), 1:10))
}
