% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{getProteinDat}
\alias{getProteinDat}
\title{getProteinDat}
\usage{
getProteinDat(gene, edb, bsg, txid = NULL, domain.src = "pfam",
  default.longest = TRUE)
}
\arguments{
\item{gene}{Gene of interest in HUGO format (e.g., MTOR)}

\item{edb}{Ensembl DB R package, v86=GRCh38, v75=GRCh37 (Default: EnsDb.Hsapiens.v86)}

\item{txid}{Primary transcript ID analyzing. Returns first TXID if none provided (Default: NULL)}

\item{domain.src}{Protein feature databse, only pfam supported at this time (Default: pfam)}

\item{default.longest}{Selects the longest ENST txid as the canonical txid (Default: TRUE)}
}
\description{
Retrieves the protein structure and features from the ensembldb database.
}
\examples{
getProteinDat("MTOR", EnsDb.Hsapiens.v86, BSgenome.Hsapiens.NCBI.GRCh38)
}
