% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRleIdx}
\alias{getRleIdx}
\title{utils: getRleIdx}
\usage{
getRleIdx(x, col.id = NA, na.val = -100)
}
\arguments{
\item{x}{A sequence to perform RLE on.  Can be either a vector or data frame}

\item{col.id}{If dataframe is provided as \code{x}, a column index is required}

\item{na.val}{Default value to set NAs too so RLE indexing doesn't break}
}
\description{
Similar to the rle() function, but gives start, end and NA indices on top of the existing "length" and "values". Easily viewed using str() on the return object().
}
\examples{
getRleIdx(c(rep(1,5), rep(2,5), rep(3,5)))
}
