% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune_Diversity.R
\name{immunelistfx}
\alias{immunelistfx}
\title{List clone counts for diversity measures}
\usage{
immunelistfx(datapath, chain)
}
\arguments{
\item{datapath}{path to mixcr files}

\item{chain}{any of: TRA, TRB, TRD, TRG, IGH, IGL, IGK}
}
\description{
This funtion lists read counts (column "cloneCount" on mixcr output)
for dowstream diversity calculations. Note that this function does not include files with only one clone.
This is just to remove the highly shallow samples which happens quite often if using RNAseq data. This should not cause any issues
when using the function with captured data.
}
\examples{
immunelistfx("~/git/PLTK/PLTK/data-raw/", "TRB")


}
