% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{initializeGrPlot}
\alias{initializeGrPlot}
\title{GR Viz: Initializes the parameters for GRanges visualization and Cytoband tracks}
\usage{
initializeGrPlot(cytoband.gr, target.chr = NULL, plot.chrom = TRUE,
  ...)
}
\arguments{
\item{cytoband.gr}{[GRanges]: A Granges object containing all UCSC cytobands, for example, PLTK::hg19.cytoband}

\item{target.chr}{[Character]: An optional target chromosome to plot.  If no chromosome is given, the entire genome is plotted [Default: NULL]}

\item{plot.chrom}{[Boolean]: Whether to generate a plot or not}

\item{...}{}
}
\description{
Main function is to initialize the adjusted x coordinates for all chromosomes but also optionally plots a track chromosome and cytoband/centromere positions
}
