% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{legend.col}
\alias{legend.col}
\title{legend.col}
\usage{
legend.col(col, lev)
}
\arguments{
\item{col}{Color ramp Palette}

\item{lev}{Levels of your data}
}
\description{
Insanely handy function to create a legend obtained from https://aurelienmadouasse.wordpress.com/2012/01/13/legend-for-a-continuous-color-scale-in-r/
}
\examples{
myDat <- data.frame(
x = runif(100),
y = runif(100),
z = rnorm(100))

myDat$Z <- cut(myDat$z, 100, label = FALSE)

colr <- rev(heat.colors(100))
plot(y ~ x, data = myDat,
     col = colr[Z], pch = 20)

legend.col(col = colr, lev = myDat$z)
}
