% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcr_QCplot.R
\name{mixcrQC.fx}
\alias{mixcrQC.fx}
\title{Plot Cap-TCRseq QC metrics}
\usage{
mixcrQC.fx(datapath, alignstatsfile, assemblestatsfile, plotname, plotpath)
}
\arguments{
\item{datapath}{path to log_parser output files}

\item{alignstatsfile}{align_stats.csv filename (from log_parser)}

\item{assemblestatsfile}{assemble_stats.csv filename (from log_parser)}

\item{plotname}{name for output plot}

\item{plotpath}{path to plot directory}
}
\description{
This function generates series of plots as QC metrics for Cap-TCRseq runs. Input files are csv files generated from log_parser.py.
Samples are ordered by total sequencing reads.
}
