% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcr_QC_multistudy.R
\name{mixcrQC.multistudy.fx}
\alias{mixcrQC.multistudy.fx}
\title{Plot Cap-TCRseq QC metrics across multiple studies}
\usage{
mixcrQC.multistudy.fx(align_assemble_stats_df, plotname, plotpath,
  sampletype)
}
\arguments{
\item{align_assemble_stats_df}{a merged file generated from align_stats and assemble_stats file.}

\item{plotname}{name for output plot}

\item{plotpath}{path to plot directory}

\item{sampletype}{type of sample eg: cfDNA, Blood or Tumor}
}
\description{
This function generates series of plots as QC metrics for Cap-TCRseq data from multiple studies.
Samples are ordered by total sequencing reads within each study.
}
