% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{... (e.g., p and q) in this case are individual plots}

\item{plotlist}{No idea?}

\item{cols}{Number of columns}

\item{layout}{Matrix specificy layout (cols and rows)}
}
\description{
As function states, plot multiple plots onto one figure
}
\examples{
multiplot(p, q) 
}
