% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{mutateCds}
\alias{mutateCds}
\title{mutateCds}
\usage{
mutateCds(Indexed_Seq, Start_Position, End_Position, Variant_Type,
  Alt_Allele)
}
\arguments{
\item{Indexed_Seq}{An Indexed and named vector of CDS sequences
Indeces are relative to transcription start, names are according to genomic
position.}

\item{Start_Position}{Mutation start position}

\item{End_Position}{Mutation end position}

\item{Variant_Type}{Variant type, options: c("SNP","INS","DEL")}

\item{Alt_Allele}{Alternate allele to be added}

\item{Chromosome}{Chromosome}
}
\description{
Given a CDS data frame and a mutation, return the Amino-Acid
sequence, aligned to the WT sequence
}
\examples{
mutateCds("ENST00000471181", EnsDb.Hsapiens.v86, BSgenome.Hsapiens.UCSC.hg38)
}
