% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonetrack.R
\name{plot_clonetracks.fx}
\alias{plot_clonetracks.fx}
\title{Plot recurring CDR3s}
\usage{
plot_clonetracks.fx(compldfle, plotpath, chain, countfrac, clnefrc)
}
\arguments{
\item{compldfle}{dataframe from cdr3_dataframe.fx}

\item{plotpath}{path to plot directory}

\item{chain}{any of: TRA, TRB, TRD, TRG, IGH, IGL, IGK}

\item{countfrac}{plot either clonal fraction of absolute counts
("cloneCount" or "cloneFraction")}

\item{clnefrc}{specify a cut-off from 0 to 1 to track and plot only a subset of clonotypes.
Useful when you have too many clonotypes to plot.
Clonal fraction of 0.001 is usually a good starting point. If you want all just pass 0.}
}
\description{
Hello!


This function uses the dataframe from cdr3_dataframe.fx function,
identifies recurring nucleotideCDR3s (found in at least two samples) and color them on clones stacked barplot. If no recurring CDR3s are found,
it outputs the clones stacked barplot with no colors.

If you need to clean up samplenames or change the order of samples on the stacked barplot,
just modify your dataframe.
}
