% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_viz.R
\name{plotGrMetadata}
\alias{plotGrMetadata}
\title{GR Viz: Main for plotting GRanges objects}
\usage{
plotGrMetadata(gr, plot.settings, data.type = "cn", target.chr = NULL,
  add.axis = TRUE, side = 1, axis.marks = 5, yrange = c(0, 1),
  y.spacer = 0.1, ...)
}
\arguments{
\item{gr}{[GRanges]: Input GRanges object}

\item{plot.settings}{[List]: Output of initializeGrPlot():  Meant to keep all GRanges plots aligned}

\item{data.type}{[Character]: Data type:  IGV Copy-number Tracks ("cn"), "expr", "mutation"}

\item{target.chr}{[Character]: Target chromosome to plot.  Must be the same one specified in initializeGrPlot()}

\item{add.axis}{[Boolean]: Whether to add an x-axis for the plot}

\item{side}{[Integer]: What side to put the axis (1=below, 3=above) [Default:1]}

\item{axis.marks}{[Integer]: How many axis tick marks to add [Default:5]}

\item{yrange}{[Integer Vector]: Your ylims for plot() function.  Everything will scale to fit within this range [Default: c(0,1)]}

\item{y.spacer}{[Numeric]: The amount of space to add between CN tracks; for example, a spacer of 0.1 for one sample in a yrange of c(0,1) will add a 0.1 gap and a 0.05 gap for 2 samples [Default: 0.1]}

\item{...}{}
}
\description{
A general wrapper for plotting GRanges objects assuming the elementMetadata() contains the numeric values of plotting.  Specifies the analysis/visualization type and it will generate a plot with the aforementioned data
}
