% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotLikelihoodRatio}
\alias{plotLikelihoodRatio}
\title{Visualization: plotLikelihoodRatio}
\usage{
plotLikelihoodRatio(groups, xmat, f1col = "black", f2col = "red",
  gen.plot = TRUE, upp.y = 2, low.y = -0.2, parmar = c(5.1, 4.1,
  4.1, 4.1), addSegments = T, col.idx = NA, ...)
}
\arguments{
\item{groups}{A list of numerical vectors (groups) or a matrix/dataframe with each column being its own group}

\item{xmat}{A 1-row matrix or vector with identification of x-values to calculate point likelihood ratios}

\item{f1col}{Group 1 colour}

\item{f2col}{Group 2 colour}

\item{gen.plot}{Generates plots}

\item{upp.y}{Scaling for plotting the y-axis; keeps all y-values proportional}

\item{low.y}{Scaling for plotting the y-axis; keeps all y-values proportional}

\item{parmar}{par()$mar values if you want to change default}

\item{addSegments}{Boolean on whether to add rectangular segments showing heights of likelihoods}

\item{col.idx}{The index of the two columns to use if given a dataframe/matrix with more than 2 columns}

\item{...}{Passed into density() and plot() function}
}
\description{
Creates a KDE curve for two datasets and calculates the likelihood ratio between the curves by calculating the natural log-ratio of point-probability estimates between the two datasets. Originally used to quantify the differences in drug-sensitivity for a given cell line (REH) between two datasets (CCLE and GDSC) for paclitaxel.
}
\examples{
demo.plotLikelihoodRatio()
}
