% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_snv.R
\name{plotProteinStructure}
\alias{plotProteinStructure}
\title{plotProteinStructure}
\usage{
plotProteinStructure(pdat, st.range, end.range, cex.val = 0.7)
}
\arguments{
\item{pdat}{Output list data structure from getProteinDat()}

\item{st.range}{Start amino acid index for region of interest}

\item{end.range}{End amino acid index for region of interest}

\item{cex.val}{CEX value for the writing}
}
\description{
Plots the top protein structure data track
}
\examples{
pdf("~/Desktop/test2.pdf", width=15)
split.screen(c(3,1))
screen(1)
plotProteinStructure(pdat)
close.screen(all.screens=TRUE)
}
