% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeFactors}
\alias{removeFactors}
\title{Converts factors to numeric or character in a data frame}
\usage{
removeFactors(df)
}
\arguments{
\item{df}{}
}
\description{
Converts factors to numeric or character in a data frame
}
