% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{runCnSignatures}
\alias{runCnSignatures}
\title{cnSignature: Wrapper to run all CN Signatures}
\usage{
runCnSignatures(gr, binsize = 1e+06, bins = PLTK::bins,
  gap = PLTK::hg19.centromeres, gap.type = "centromeres",
  assign.amp.del = FALSE, cn.thresh = 0.5, cn.scale = 0, ...)
}
\arguments{
\item{gr}{[GRanges]: GRanges object with copy-number in the elementMetadata().  Tested on output from convertToGr()}

\item{binsize}{[Integer]: Maximum size of segments to be considered in a small-CN cluster [Default: 1000000]}

\item{bins}{[GRanges]: A pre-made GRanges object where the genome is binned into smaller segment [Default: PLTK::bins]}

\item{gap}{}

\item{gap.type}{}

\item{assign.amp.del}{}

\item{cn.thresh}{}

\item{cn.scale}{}

\item{...}{}
}
\value{
[List]: List of all signatures List[[each sample]][[Signature1]]
}
\description{
A Wrapper function to run all copy-number GRanges objects through the CN signature functions
}
