% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{segfileToGr}
\alias{segfileToGr}
\title{cnTools: Converts the .seg file to a GRanges object}
\usage{
segfileToGr(seg, col.id = "SampleX")
}
\arguments{
\item{seg}{A dataframe containing the seg file, with headers}

\item{col.id}{Name of sample [Default: SampleX]}
}
\value{
A Granges object with one column in the element Metadata() corresponding to the sample
}
\description{
Uses the standard seg inputs (https://software.broadinstitute.org/software/igv/sites/cancerinformatics.org.igv/files/linked_files/example.seg) and converts it into a list of GRanges object with the seg.mean column being the copy-number data stored in elementMetadata()
}
