% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{sigBinBreakpoints}
\alias{sigBinBreakpoints}
\title{cnSignature: sigBinBreakpoints}
\usage{
sigBinBreakpoints(gr, bins, numeric.return = FALSE)
}
\arguments{
\item{gr}{GRanges object of your segments}

\item{bins}{GRanges object of the genome binned into set segment sizes [i.e. PLTK::bins]}
}
\value{
A list containing two elements:
  \code{segs}: A list of granges object for the original \code{gr} placed into each \code{bins}
  \code{bins}: The original \code{bins} object with an additional column in the metadata indicating number of breakpoints
}
\description{
Takes a GRanges object of segments and counts the number of breakpoints found within pre-designed genomic bins
}
\examples{
 sigBinBreakpoints(PLTK::genDemoData(), PLTK::bins)
}
