% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{sigClusterBreakpoints}
\alias{sigClusterBreakpoints}
\title{cnSignature: sigClusterBreakpoints}
\usage{
sigClusterBreakpoints(gr, binsize, numeric.return = FALSE)
}
\arguments{
\item{gr}{A granges object}

\item{binsize}{A set binsize (bp)}

\item{numeric.return}{}
}
\value{
\code{segs.list}: A list of genomicRanges objects for genomic regions with continuous segments smaller than binsize
}
\description{
Takes a list of copy-number segments and tries to identify regions where there are consecutive segments less than a pre-designed segment size. Within these regions, it finds the longest string of consecutive segments that are less than the pre-designed segment size, annotates it, and reports them back in a list for downstream analysis.
}
\examples{
sigClusterBreakpoints(PLTK::genDemoData(), 50)
}
