% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{sigCnChangepoint}
\alias{sigCnChangepoint}
\title{cnSignature: Copy-number changepoints}
\usage{
sigCnChangepoint(gr, collapse.segs = FALSE)
}
\arguments{
\item{gr}{[GRanges]: GRanges object}

\item{collapse.segs}{[Boolean]: Whether segments with same copy-states separated by a gap with no data should be collapsed}
}
\value{
List containing and element for each chromosome with the copy-number changepoints at each breakpoint
}
\description{
Calculates the copy-number change at breakpoint per chromosome
}
\examples{
sigCnChangepoint(gr, collapse.segs=FALSE)
}
