% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{sigCopyNumber}
\alias{sigCopyNumber}
\title{cnSignature: All copy numbers}
\usage{
sigCopyNumber(gr, weight = TRUE, normalize = FALSE)
}
\arguments{
\item{gr}{[GRanges]: GRanges object}

\item{weight}{[Boolean]: Weight the copy-number for a segment based on the size of the segment}

\item{normalize}{[Boolean]: Normalize by chromosome lengths; uses PLTK::hg19.cytobands as default chromosome sizes}
}
\value{
Dataframe of copy-number values for each chromosome
}
\description{
Calculates the copy numbers in the sample copy-number data with the option to scale it based on the size of the segment
}
\examples{
sigCopyNumber(gr, weight=TRUE, normalize=TRUE)
}
