% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{sigGapDist}
\alias{sigGapDist}
\title{cnSignature: Distance from Centromeres or Telomeres}
\usage{
sigGapDist(gr, gap = PLTK::hg19.centromeres, gap.type = "centromeres",
  normalize = FALSE, verbose = FALSE, numeric.return = FALSE)
}
\arguments{
\item{gr}{[GRanges]: GRanges object}

\item{gap}{[GRanges]: GRanges gap data, either PLTK::centromeres or PLTK::telomeres}

\item{gap.type}{[Character]: Either "centromeres" or "telomeres"}

\item{normalize}{[Boolean]: Normalize by chromosome lengths; uses PLTK::hg19.cytobands as default chromosome sizes}

\item{verbose}{[Boolean]: Prints out extra debug info}

\item{numeric.return}{}
}
\value{
List of distances from the telomere or centromere
}
\description{
Calculates the distances from the closest breakpoint end to the telomere or centromere.  Offers the option to normalize by the size of the chromosome.
}
\examples{
sigGapDist(demo, gap.type = "telomeres", gap = PLTK::hg19.telomeres, normalize=TRUE)
}
