% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeNumColumns}
\alias{summarizeNumColumns}
\title{Summarizes all numeric values in a dataframe}
\usage{
summarizeNumColumns(df)
}
\arguments{
\item{df}{A dataframe to try and summarize}
}
\description{
Summarizes all numeric values in a dataframe
}
