% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSignature.R
\name{summarizeSignatures}
\alias{summarizeSignatures}
\title{cnSignature: Summarizes the CN signature list}
\usage{
summarizeSignatures(sig, ids = NULL, decompose = TRUE,
  sig.metric.values = c("mean", "sd", "skew", "kurtosis", "modality"))
}
\arguments{
\item{sig}{[List]: The returned value from runCnSignatures() with numeric.return=TRUE}

\item{ids}{[Char. Vector]: Vector of all sample names}

\item{sig.metric.values}{[Char. Vector]: All values to be returned from describe and mclust}
}
\description{
Works with the return list from runCnSignatures() when numeric.return=TRUE.  Will summarize this into a matrix
}
