% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot publication theme}
\usage{
theme_Publication(base_size = 18, base_family = "helvetica")
}
\arguments{
\item{base_size}{Default=18}

\item{base_family}{Default='helvetica'}
}
\value{
Theme for ggplot
}
\description{
ggplot theme for publication ready Plots: https://rpubs.com/Koundy/71792
}
\examples{
data <- data.frame("A"=letters[1:10], "Val"=1:10)
ggplot(x, aes(x=A, y=Val)) +
geom_point() + 
theme_Publication()
}
