% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{annotateSegments}
\alias{annotateSegments}
\title{cnTools: annotateSegments}
\usage{
annotateSegments(cn.data, genes)
}
\arguments{
\item{cn.data}{A dataframe that can be converted to GRanges object easily, or a granges object}

\item{genes}{A GRanges object of genes with gene_ids housing annotation data. Easiest as the output from getGenes()}
}
\value{
Annotated GRanges object with gene ids for the input GRanges
}
\description{
cnTools: annotateSegments
}
\examples{
annotateSegments(PLTK::genDemoData(), PLTK::getGenes())
}
