% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{assignAmpDel}
\alias{assignAmpDel}
\title{cnTools: Convert Log2Ratio to Amp/Del}
\usage{
assignAmpDel(seg.gr, cn.thresh = 0.5, cn.scale = 0)
}
\arguments{
\item{seg.gr}{[GRanges]:  A GRanges object with copy-number log2ratios in elementMetadata() columns}

\item{cn.thresh}{[Integer]: A log2ratio cutoff to indicate gain or loss}

\item{cn.scale}{[Numeric]: X scale}
}
\value{

}
\description{
A temporary helper function to truncate log2ratios at 0.5 and turn them into Amp (1) or Del (-1)
}
