% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{cnGenomeFraction}
\alias{cnGenomeFraction}
\title{cnTools: Wrapper for copy-number metrics}
\usage{
cnGenomeFraction(analysis, gr, cn.stat = "all", ...)
}
\arguments{
\item{analysis}{[Character]: passed in from cnMetrics}

\item{gr}{[GRanges]: passed in from cnMetrics}

\item{...}{}

\item{[Character]:}{cn.stat passed in from cnMetrics}
}
\value{

}
\description{
Calculates Genomic Fraction or wGII scores. Refer to cnMetrics for more detail.
}
