% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{cnGenomeFraction}
\alias{cnGenomeFraction}
\title{cnTools: Wrapper for copy-number metrics}
\usage{
cnGenomeFraction(analysis, gr, cn.stat = "all", copy.neutral,
  cn.variance)
}
\arguments{
\item{analysis}{}

\item{gr}{}

\item{cn.stat}{}

\item{copy.neutral}{}

\item{cn.variance}{}

\item{...}{analysis, gr pass in from cnMetrics}
}
\description{
Calculates Genomic Fraction or wGII scores. Refer to cnMetrics for more detail.
}
