% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{cnMetrics}
\alias{cnMetrics}
\title{cnTools: Wrapper for copy-number metrics}
\usage{
cnMetrics(analysis = NA, gr, cn.stat = "all", copy.neutral = 0)
}
\arguments{
\item{[GRanges]:}{gr GRanges object}

\item{[Character]:}{cn.stat 'all' for all CN-aberrations or 'gain' or 'loss'}

\item{[Integer]:}{copy.neutral Integer specifying what a copy-neutral value is [default = 0]}

\item{[Character]:}{analysis The analysis to perform: "gf" genomic fraction, "wgii" for wGII scores (genomic fraction normalized for chromosome)}
}
\description{
A wrapper to run copy-number analysis on a GRanges copy-number dataset with copy-number values stored in elementMetadata().  These metrics include calculating genomic fraction and wGII scores (bp) for gains, losses, or any CN-abberation.
}
