% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{mapGrToReference}
\alias{mapGrToReference}
\title{cnTools: Maps one GRanges objec to another (Intersect)}
\usage{
mapGrToReference(gr, ref.gr, overlap = "mode", mode.type = "normal")
}
\arguments{
\item{gr}{[GRanges]: Input GRanges object with CNdata stored in elementMetadata()}

\item{ref.gr}{[GRanges]: Target GRanges object to map to}

\item{overlap}{[Character]: How to handle when multiple ranges from input_GRanges are within a single bin of the target_GRanges.  Options: "mode", "mean", "null" (all are weighted by segment sizes)}

\item{mode.type}{[Character]: "normal" or  "quick"}
}
\value{
The same Target GRanges object with elementMetadata() filled in
}
\description{
cnTools: Maps one GRanges objec to another (Intersect)
}
