% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnTools.R
\name{overlapGrToReference}
\alias{overlapGrToReference}
\title{Title}
\usage{
overlapGrToReference(gr, ref.gr, int_ids = c("start", "end"),
  chr_id = "chr", overlap)
}
\arguments{
\item{gr}{}

\item{ref.gr}{}

\item{int_ids}{}

\item{chr_id}{}

\item{overlap}{}
}
\description{
Title
}
